+++ src/pages/incident/incident.ts
@@ -207,7 +207,7 @@ export default class IncidentPage extends Vue {
   public async getStreamsData (): Promise<void> {
     const params: string = this.$route.params.projectId as string
     if (!params) return
-    const streamsData = await StreamService.getStreams({ projects: [params] })
+    const streamsData = await StreamService.getStreams({ projects: [params], limit: 1000 })
     this.streamsData = streamsData.data
   }