+++ config/index.ts
@@ -1,3 +1,3 @@
-const rangerBaseApi: string = import.meta.env.MODE === 'production' ? 'https://ranger-api.rfcx.org' : 'https://staging-ranger-api.rfcx.org'
+const rangerBaseApi: string = import.meta.env.MODE === 'production' ? 'https://guardian-api.rfcx.org' : 'https://staging-guardian-api.rfcx.org'
 
 export default rangerBaseApi