+++ src/components/pagination/pagination.vue
@@ -9,24 +9,24 @@
         class="cursor-pointer relative inline-flex items-center px-4 py-2 border border-gray-500 text-sm font-medium rounded-md text-white bg-gray hover:bg-gray-700"
         @click="getPrevPage()"
       >
-        Previous
+        {{ $t('Previous') }}
       </a>
       <a
         :class="{'opacity-50 cursor-not-allowed': paginationSettings.page === lastPage}"
         class="cursor-pointer ml-3 relative inline-flex items-center px-4 py-2 border border-gray-500 text-sm font-medium rounded-md text-white bg-gray hover:bg-gray-700"
         @click="getNextPage()"
       >
-        Next
+        {{ $t('Next') }}
       </a>
     </div>
     <div class="hidden sm:flex-1 sm:flex sm:items-center sm:justify-between">
       <div>
         <p class="text-sm text-white">
-          Page
+          {{ $t('Page') }}
           {{ ' ' }}
           <span class="font-medium">{{ paginationSettings.page }}</span>
           {{ ' ' }}
-          of
+          {{ $t('of') }}
           {{ ' ' }}
           <span class="font-medium">{{ lastPage }}</span>
         </p>
@@ -41,7 +41,7 @@
             class="relative inline-flex items-center px-2 py-2 rounded-l-md border border-gray-500 bg-gray text-sm font-medium text-white hover:bg-gray-700"
             @click="getPrevPage()"
           >
-            <span class="sr-only">Previous</span>
+            <span class="sr-only">{{ $t('Previous') }}</span>
             <ChevronLeftIcon
               class="h-5 w-5"
               aria-hidden="true"
@@ -62,7 +62,7 @@
             class="relative inline-flex items-center px-2 py-2 rounded-r-md border border-gray-500 bg-gray text-sm font-medium text-white hover:bg-gray-700"
             @click="getNextPage()"
           >
-            <span class="sr-only">Next</span>
+            <span class="sr-only">{{ $t('Next') }}</span>
             <ChevronRightIcon
               class="h-5 w-5"
               aria-hidden="true"